requirements.txt
apscheduler=3.11.0
asyncgui==0.6.3
asynckivy==0.6.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
docutils==0.21.2
filetype==1.2.0
greenlet==3.1.1
idna==3.10
Kivy==2.3.1
Kivy-Garden==0.1.5
materialyoucolor==2.0.10
pillow==11.1.0
plyer=2.1.0
pycparser==2.22
Pygments==2.19.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.38
typing_extensions==4.13.1
urllib3==2.3.0
